part3/requirements.txt
flask
flask-restx
flask-sqlalchemy
flask-jwt-extended
flask-bcrypt
marshmallow
